start the users api standardization and documentation
